requirements.txt
amqp==2.4.2
asn1crypto==0.24.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.6.1
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.5.0
MarkupSafe==1.1.1
optionaldict==0.1.1
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
requests==2.21.0
six==1.12.0
some-package==0.1
urllib3==1.24.3
vine==1.3.0
wechatpy==1.8.2
Werkzeug==0.15.2
xmltodict==0.12.0
